+++ src/main/java/in/gov/kfon/dmdm/repository/DfBankDetailsRepository.java
@@ -1,0 +1,7 @@
+package in.gov.kfon.dmdm.repository;
+
+import in.gov.kfon.dmdm.model.DfBankDetails;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DfBankDetailsRepository extends JpaRepository<DfBankDetails, UUID> {}