+++ src/main/java/in/gov/kfon/dmdm/repository/DfTransDetailsRepository.java
@@ -1,0 +1,7 @@
+package in.gov.kfon.dmdm.repository;
+
+import in.gov.kfon.dmdm.model.DfTransDetails;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DfTransDetailsRepository extends JpaRepository<DfTransDetails, UUID> {}